added smss class

+++ src/SMSS_Problem.java
@@ -1,6 +1,6 @@
 public class SMSS_Problem {
     public int[] sequence;
-    public long max_sum;
+    public int max_sum;
     public SMSS_Problem(int[] sequence) {
         this.sequence = sequence;
         this.max_sum = 0;
